sixth commit

+++ backend/foodgram/api/views.py
@@ -22,14 +22,9 @@ from api.serializers import (
     RecipesWriteSerializer,
     TagsSerializer,
 )
-from recipes.models import (Favorite,
-			    Ingredient,
-			    Recipe,
-			    ShoppingCart,
-			    Tag)
+from recipes.models import Favorite, Ingredient, Recipe, ShoppingCart, Tag
 
 User = get_user_model()
-
 
 class RecipesViewSet(viewsets.ModelViewSet):
     queryset = Recipe.objects.all()
@@ -47,20 +42,17 @@ class RecipesViewSet(viewsets.ModelViewSet):
         queryset = Recipe.objects.all()
         author = self.request.user
         if self.request.GET.get('is_favorited'):
-            favorite_recipes_ids = Favorite.objects \
-            			   .filter(user=author).values('recipe_id')
+            favorite_recipes_ids = Favorite.objects.filter(user=author).values('recipe_id')
             return queryset.filter(pk__in=favorite_recipes_ids)
 
         if self.request.GET.get('is_in_shopping_cart'):
-            cart_recipes_ids = ShoppingCart.objects \
-            		       .filter(user=author).values('recipe_id')
+            cart_recipes_ids = ShoppingCart.objects.filter(user=author).values('recipe_id')
             return queryset.filter(pk__in=cart_recipes_ids)
         return queryset
 
     def add_in_list(self, model, user, pk):
         if model.objects.filter(user=user, recipe__id=pk).exists():
-            return Response({'errors': f'Рецепт уже добавлен в {model.__name__}'},
-            		    status=status.HTTP_400_BAD_REQUEST)
+            return Response({'errors': f'Рецепт уже добавлен в {model.__name__}'}, status=status.HTTP_400_BAD_REQUEST)
         recipe = get_object_or_404(Recipe, pk=pk)
         model.objects.create(user=user, recipe=recipe)
         serializer = RecipeListSerializer(recipe)
@@ -71,25 +63,21 @@ class RecipesViewSet(viewsets.ModelViewSet):
         if obj.exists():
             obj.delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
-        return Response({'errors': f'Рецепт не добавлен в {model.__name__}'},
-                        status=status.HTTP_400_BAD_REQUEST)
+        return Response({'errors': f'Рецепт не добавлен в {model.__name__}'}, status=status.HTTP_400_BAD_REQUEST)
 
-    @action(methods=['post', 'delete'], detail=True,
-    	    permission_classes=[IsAuthenticated])
+    @action(methods=['post', 'delete'], detail=True, permission_classes=[IsAuthenticated])
     def favorite(self, request, pk=None):
         if request.method == 'POST':
             return self.add_in_list(Favorite, request.user, pk)
         return self.delete_in_list(Favorite, request.user, pk)
 
-    @action(methods=['post', 'delete'],
-    	    detail=True, permission_classes=[IsAuthenticated])
+    @action(methods=['post', 'delete'], detail=True, permission_classes=[IsAuthenticated])
     def shopping_cart(self, request, pk=None):
         if request.method == 'POST':
             return self.add_in_list(ShoppingCart, request.user, pk)
         return self.delete_in_list(ShoppingCart, request.user, pk)
 
-    @action(methods=['GET'], detail=False,
-    	    permission_classes=[IsAuthenticated])
+    @action(methods=['GET'], detail=False, permission_classes=[IsAuthenticated])
     def download_shopping_cart(self, request):
         ingredients = request.user.shopping_cart.values(
             'recipe__ingredients_amount__ingredient__name',
@@ -112,12 +100,10 @@ class RecipesViewSet(viewsets.ModelViewSet):
         )
         return response
 
-
 class TagsViewSet(viewsets.ReadOnlyModelViewSet):
     queryset = Tag.objects.all()
     serializer_class = TagsSerializer
     pagination_class = None
-
 
 class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
     serializer_class = IngredientSerializer
@@ -127,35 +113,22 @@ class IngredientViewSet(viewsets.ReadOnlyModelViewSet):
     search_fields = ['^name']
     pagination_class = None
 
-
 class FollowUserView(APIView):
     permission_classes = [IsAuthenticated]
 
     def post(self, request, id):
         author = get_object_or_404(User, id=id)
         if request.user.follower.filter(author=author).exists():
-            return Response(
-                {"errors": "Вы уже подписаны на автора"},
-                status=status.HTTP_400_BAD_REQUEST,
-            )
-        serializer = FollowSerializer(
-            request.user.follower.create(author=author),
-            context={"request": request},
-        )
+            return Response({"errors": "Вы уже подписаны на автора"}, status=status.HTTP_400_BAD_REQUEST)
+        serializer = FollowSerializer(request.user.follower.create(author=author), context={"request": request})
         return Response(serializer.data, status=status.HTTP_201_CREATED)
 
     def delete(self, request, id):
         author = get_object_or_404(User, id=id)
         if request.user.follower.filter(author=author).exists():
-            request.user.follower.filter(
-                author=author
-            ).delete()
+            request.user.follower.filter(author=author).delete()
             return Response(status=status.HTTP_204_NO_CONTENT)
-        return Response(
-            {"errors": "Автор отсутствует в списке подписок"},
-            status=status.HTTP_400_BAD_REQUEST,
-        )
-
+        return Response({"errors": "Автор отсутствует в списке подписок"}, status=status.HTTP_400_BAD_REQUEST)
 
 class SubscriptionsView(ListAPIView):
     serializer_class = FollowSerializer
